Normalize playback volume, and fix the gain at gapless joins
The normalization switch has been in Settings since before the engine existed
and has never moved a sample: `QueueEntry.gain_db` reached the mixer correctly,
but nothing ever put a number in it.

Gains now come from two places. Most libraries already carry ReplayGain tags,
which lofty exposes on the same keys whatever the container, so reading them is
free. Files without them — and WAV and AIFF have nowhere to put them — are
measured against EBU R128 *while they play*: the engine is decoding those
samples anyway, so a track heard end to end records its own loudness for next
time and the library fills in as it is used. A seek abandons the measurement
rather than storing a partial listen, and silence or anything under R128's
400 ms gate reports no result at all — left unguarded that is an infinite gain,
so it is guarded twice.

Reading the tags needs a tolerant parser. Taggers write `-6.94 dB`, `+3.21 dB`,
`-6.94dB`, a bare `-6.94`, and, from a European locale, `-6,94 dB`. The comma
is the one that silently defeats a naive parser on a real library, so it has a
test.

This exposed a bug that was invisible while every gain was 1.0. At a gapless
join the engine set the incoming track's gain immediately, while up to half a
second of the outgoing track was still in the ring — so each track's tail would
have played at the next one's volume. The gain now rides on the existing
`Boundary` and is applied when the join actually reaches the device, which is
already how `TrackChanged` is timed. `Gain` ramps across the buffer, so the
change is inaudible.

The switch is enforced in the engine rather than by sending zero from the
frontend, so turning it off is heard at once instead of at the next track.

One structural change fell out of making this testable: the engine took an
`AppHandle` to reach the database, which meant its tests needed a Tauri app.
It now depends on a two-method `loudness::Store` instead — the only thing it
actually wanted — so the audio engine has no Tauri dependency at all.

Verified against physics rather than a fixture: a full-scale mono sine measures
-3.7 LUFS, and the same tone in stereo reads 3 dB hotter, which is R128 summing
channels rather than averaging them.

+++ src-tauri/src/audio/queue.rs
@@ -19,7 +19,7 @@ pub struct QueueEntry {
     pub path: PathBuf,
     pub duration_secs: f64,
     /// Normalization gain in dB, resolved when the queue was loaded.
-    pub gain_db: f32,
+    pub gain_db: f64,
 }
 
 pub struct Queue {
